Refactor BaseEntityFindQueryToBaseEntityFindQueryTypeOrmConverterAsync
